+++ packages/graphql/tests/performance/schema/schema-performance.ts
@@ -19,7 +19,7 @@
 
 import Neo4jGraphQL from "../../../src/classes/Neo4jGraphQL";
 
-const basicTypeDefs = `
+const basicTypeDefs = /* GraphQL */ `
     type Journalist @node {
         articles: [Article!]! @relationship(type: "HAS_ARTICLE", direction: OUT, properties: "HasArticle")
     }
@@ -29,12 +29,12 @@ const basicTypeDefs = `
     }
 
     type Article @authorization(filter: [{ where: { node: { id: { eq: "$jwt.sub" } } } }]) @node {
-        id: ID! @id  @authorization(filter: [{ where: { node: { id: { eq: "$jwt.sub" } } } }])
+        id: ID! @id @authorization(filter: [{ where: { node: { id: { eq: "$jwt.sub" } } } }])
         blocks: [Block!]! @relationship(type: "HAS_BLOCK", direction: OUT, properties: "HasBlock")
         images: [Image!]! @relationship(type: "HAS_IMAGE", direction: OUT)
     }
 
-    type HasBlock @relationshipProperties @node {
+    type HasBlock @relationshipProperties {
         order: Int!
     }